+++ src/archivey/api/exceptions.py
