+++ src/archivey/formats/single_file_reader.py
@@ -5,9 +5,7 @@ import struct
 from datetime import datetime, timezone
 from typing import BinaryIO, Iterator, List
 
-from archivey.readers.base_reader import BaseArchiveReader
-from archivey.formats.compressed_streams import open_stream
-from archivey.exceptions import (
+from archivey.api.exceptions import (
     ArchiveFormatError,
 )
 from archivey.api.types import (
@@ -18,6 +16,8 @@ from archivey.api.types import (
     CreateSystem,
     MemberType,
 )
+from archivey.formats.compressed_streams import open_stream
+from archivey.internal.base_reader import BaseArchiveReader
 
 logger = logging.getLogger(__name__)
