+++ src/archivey/internal/extraction_helper.py
@@ -8,11 +8,14 @@ import threading
 from typing import TYPE_CHECKING, BinaryIO
 
 from archivey.api.config import OverwriteMode
-from archivey.exceptions import ArchiveFileExistsError, ArchiveLinkTargetNotFoundError
+from archivey.api.exceptions import (
+    ArchiveFileExistsError,
+    ArchiveLinkTargetNotFoundError,
+)
 from archivey.api.types import ArchiveMember, MemberType
 
 if TYPE_CHECKING:
-    from archivey.readers.base_reader import ArchiveReader
+    from archivey.internal.base_reader import ArchiveReader
 
 logger = logging.getLogger(__name__)
