+++ tests/archivey/testing_utils.py
@@ -9,12 +9,12 @@ from typing import Optional
 import pytest
 
 from archivey.api.config import ArchiveyConfig
-from archivey.dependency_checker import get_dependency_versions
 from archivey.api.types import (
     TAR_FORMAT_TO_COMPRESSION_FORMAT,
     ArchiveFormat,
     MemberType,
 )
+from archivey.internal.dependency_checker import get_dependency_versions
 from tests.archivey.sample_archives import (
     FileInfo,
 )
