+++ tests/create_archives.py
@@ -56,8 +56,12 @@ try:  # Optional dependency
 except ModuleNotFoundError:
     zstandard = None
 
-from archivey.exceptions import PackageNotInstalledError
-from archivey.api.types import TAR_FORMAT_TO_COMPRESSION_FORMAT, ArchiveFormat, MemberType
+from archivey.api.exceptions import PackageNotInstalledError
+from archivey.api.types import (
+    TAR_FORMAT_TO_COMPRESSION_FORMAT,
+    ArchiveFormat,
+    MemberType,
+)
 from tests.archivey.sample_archives import (
     SAMPLE_ARCHIVES,
     ArchiveContents,
